+++ python-backend/app/api/services/NRIS_service.py
